Fix : 갤러리 이동 메서드 MainActiviy로, rebase 오류 해결
